fix(lint): remove unused import and use destructuring for getAvailablePods
