+++ src/api.v2/helpers/worker/workSubmit/createWorkerK8s.js
@@ -1,6 +1,5 @@
 const k8s = require('@kubernetes/client-node');
 const config = require('../../../../config');
-const asyncTimer = require('../../../../utils/asyncTimer');
 const getLogger = require('../../../../utils/getLogger');
 
 const kc = new k8s.KubeConfig();
@@ -53,17 +52,8 @@ const getDeployment = async (name, namespace) => {
   return deployment;
 };
 
-const waitForPods = async (namespace, maxTries = 20, currentTry = 0) => {
-  logger.log('Waiting for pods...');
-  await asyncTimer(1000);
-  const pods = await getAvailablePods(namespace);
 
-  if (pods.length > 0 || currentTry >= maxTries) {
-    return;
-  }
-
-  await waitForPods(namespace, maxTries, currentTry + 1);
-};
+const waitForPods = require('./waitForPods');
 
 
 const scaleDeploymentReplicas = async (name, namespace, deployment, desiredReplicas = 1) => {
@@ -153,3 +143,4 @@ const createWorkerResources = async (service) => {
 };
 
 module.exports = createWorkerResources;
+module.exports.getAvailablePods = getAvailablePods;
