+++ src/api.v2/helpers/worker/workSubmit/waitForPods.js
@@ -1,0 +1,19 @@
+const asyncTimer = require('../../../../utils/asyncTimer');
+const getLogger = require('../../../../utils/getLogger');
+const { getAvailablePods } = require('./createWorkerK8s');
+
+const logger = getLogger();
+
+const waitForPods = async (namespace, maxTries = 20, currentTry = 0) => {
+  logger.log('Waiting for pods...');
+  await asyncTimer(1000);
+  const pods = await getAvailablePods(namespace);
+
+  if (pods.length > 0 || currentTry >= maxTries) {
+    return;
+  }
+
+  await waitForPods(namespace, maxTries, currentTry + 1);
+};
+
+module.exports = waitForPods;
